agentified/requirements.txt
# Core A2A and AgentBeats dependencies
a2a-sdk[http-server]>=0.3.8
earthshaker>=0.2.0

# OpenHands (for OpenAgentSafety evaluation)
# Temporarily commented out due to e2b dependency conflicts
# openhands-ai>=0.9.0

# CLI and utilities
typer>=0.19.2
uvicorn>=0.37.0
python-dotenv>=1.0.0

# LLM support
litellm>=1.0.0
openai>=1.0.0

# HTTP client
httpx>=0.27.0

# Docker control (task containers)
docker>=7.0.0

# Optional: For deployment
# google-cloud-buildpacks



